+++ lib/liquid/tags/for.rb
@@ -191,6 +191,12 @@ module Liquid
     def render_else(context)
       @else_block ? @else_block.render(context) : ''.freeze
     end
+
+    class Traversal < Liquid::Traversal
+      def children
+        (super + [@node.limit, @node.from, @node.collection_name]).compact
+      end
+    end
   end
 
   Template.register_tag('for'.freeze, For)
